requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
SQLAlchemy==0.8.3
Werkzeug==0.9.4
argparse==1.2.1
itsdangerous==0.23
path.py==4.4
requests==2.0.1
sh==1.09
slugify==0.0.1
wsgiref==0.1.2
